Added a WithConnection container that handles the communication with and from the server.
WithConnection is a container that doesn't define any graphic, its role is just that to connect with the server and handle the state of the application.
It requires a single property `render` that is the function what will actually do the rendering of the ui. This function will receive all the information it should need to draw the ui and handle the user events.